function/requirements.txt
google-cloud-storage==2.7.0
google-cloud-bigquery==3.7.0
google-cloud-secret-manager==2.12.6
pyyaml==6.0
langchain==0.0.184
functions-framework==3.3.0
langchain-openai 